[all] Killed a bunch of warnings
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@5121 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux/common/ADM_audiocodec/ADM_pluginLoad.cpp
@@ -149,7 +149,7 @@ uint8_t ADM_ad_loadPlugins(const char *path)
 	for(int i=0;i<nbFile;i++)
 		tryLoadingAudioPlugin(files[i]);
 
-	printf("[ADM_ad_plugin] Scanning done, found %u codec\n", ADM_audioPlugins.size());
+	printf("[ADM_ad_plugin] Scanning done, found %d codec\n", (int)ADM_audioPlugins.size());
 
 	return 1;
 }
